Some boss improvements
